// frontend/src/app/modules/page/digital-twin-part/model/digitalTwinPart.model.ts
/********************************************************************************
 * Copyright (c) 2025 Contributors to the Eclipse Foundation
 *
 * See the NOTICE file(s) distributed with this work for additional
 * information regarding copyright ownership.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Apache License, Version 2.0 which is available at
 * https://www.apache.org/licenses/LICENSE-2.0.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 ********************************************************************************/


import { PaginationResponse } from "@core/model/pagination.model";
import { DigitalTwinType } from "./digitalTwinType.enum";
import { DigitalTwinPartFilter } from "@shared/model/filter.model";

export interface DigitalTwinPartRequest {
  page: number;
  size: number;
  sort: string[];
  filters: DigitalTwinPartFilter[];
}

export interface DigitalTwinPartResponse {
    aasId: string;
    aasExpirationDate: Date;
    globalAssetId: string;
    assetExpirationDate: Date;
    bpn: string;
    digitalTwinType: DigitalTwinType;
  }

  export type DigitalTwinPartsResponse = PaginationResponse<DigitalTwinPartResponse>;

  export interface DigitalTwinPartDetailResponse {
    aasId: string;
    aasTTL: number;
    aasExpirationDate: Date;
    nextLookup: Date;
    globalAssetId: string;
    assetTTL: number;
    assetExpirationDate: Date;
    nextSync: Date;
    bpn: string;
    digitalTwinType: DigitalTwinType;
    actor: string
  }